ルーティング整理と型のモジュール化
