--- src/app/maps/page.tsx
@@ -1,9 +1,0 @@
-'use strict';
-
-import { redirect } from 'next/navigation';
-
-const page = async () => {
-  redirect(`/maps/1`);
-};
-
-export default page;
